proper rounding 2

+++ front/pages/api/w/[wId]/workspace-analytics.ts
@@ -148,9 +148,9 @@ async function getAnalytics(
       )
       SELECT
         "last_7_days_active_users",
-        ROUND(wow_growth_pct, 2) AS "wow_growth_pct",
+        ROUND(wow_growth_pct::Decimal, 2) AS "wow_growth_pct",
         "last_30_days_active_users",
-        ROUND(mom_growth_pct, 2) AS "mom_growth_pct"
+        ROUND(mom_growth_pct::Decimal, 2) AS "mom_growth_pct"
       FROM growth_calculations;
       `,
         {
@@ -193,8 +193,8 @@ async function getAnalytics(
           FROM averages
         )
         SELECT
-          ROUND(COALESCE(current_avg_dau, 0), 2) AS last_7_days_average_daily_active_users,
-          ROUND(COALESCE(wow_growth_pct, 0), 2) AS wow_growth_pct
+          ROUND(COALESCE(current_avg_dau, 0)::Decimal, 2) AS last_7_days_average_daily_active_users,
+          ROUND(COALESCE(wow_growth_pct, 0)::Decimal, 2) AS wow_growth_pct
         FROM wow_growth;
       `,
         {
